Fixing format strings to remove 'z' size specifier.
As of the 2013 version VC++ still doesn't support the 'z' size specifier. This makes portable printing of size_t types frustrating. The simplest general solution is to use %u and cast to unsigned. If there was any possibility of the numbers getting larger than 32-bit then we would need better alternatives, but there is not.

This was found through code inspection, through /analyze, and through pdfium_test print this non-helpful message:

Loaded, parsed and rendered zu pages.
Skipped zu bad pages.

I can confirm that the fix works on Windows and it should work identically on mac. This is a follow-on to change 02e6ca4c4f.

R=tsepez@chromium.org

Review URL: https://codereview.chromium.org/738433003