Initialize V8 external data (XFA version).
This is a quick patch to fix the segv which occurs in pdfium_test when
the v8 external data is not provided.  Fixing the full-up initialization
is left as a follow-on exercise.

Path manipulations remain a nuisance since we don't want to depend on a
particular /base library.

BUG=439793
R=thestig@chromium.org

Review URL: https://codereview.chromium.org/810793004