Fix one remaining misuse of FX_WSTRC on FX_WCHAR* vars.
FX_WSTRC is only valid on arrays, not pointers. In this code it was
being passed a pointer to a URL. This was found when integrating a
change to FX_WSTRC that disallows pointer arguments.

The consequence of this bug is that the URL will end up with incorrect
lengths. It will be one character long in 32-bit builds, and three
characters long in 64-bit builds (sizeof(WCHAR*)-1).

R=bo_xu@foxitsoftware.com

Review URL: https://codereview.chromium.org/831293002