Fix build of pdfium_embeddertest under V8_USE_EXTERNAL_STARTUP_DATA.
Currently, this is a difference between the standalone pdfium environment
vs. pdfium as part of a chromium checkout.

Locating the external data files is a nusiance when you don't have the
binary's argv[0] line, so we create a new main that preserves this for
us.

R=thestig@chromium.org

Review URL: https://codereview.chromium.org/851283006