Add output scale factor command line parameter to pdfium_test.
It turns out that pdfium rendering takes some significantly different paths
when the output bitmap size is the same as the document size, since it can
avoid work in that case. For example, to reproduce the referenced bug, a scale
factor less than 1.0 is required, so we add a parameter to let pdfium_test
cover that case.

BUG=451265
R=thestig@chromium.org

Review URL: https://codereview.chromium.org/861203003