Replace ugly JS_Define macros with templates.
This allows us to step through the JS bindings code with the
debugger, which I could not do (but wanted to) the other day.

In the process, get rid of some else after returns, and one
unreachable code path.

I also get rid of some tracing macros that we would never
use for the sake of clarity, and some plain unused
definitions.

R=brucedawson@chromium.org

Review URL: https://codereview.chromium.org/908033002