Implement a delegate for EmbedderTests.
This is the first step in allowing an embedder test to
someday gMock its callbacks, so that it can check that they
fired as expected. gMock wants a class, not a C-style
function-based API, and EmbedderTest is made to bridge
between the two.

The EmbedderTest class itself is modified to inherit from
the C JS API classes themselves, to make finding the
delegate easier.

For example, a future embedder test might send a keystroke
to a page, which would then trigger JS, which would then
trigger an Alert().  Mocking the Alert() callback would
allow the test to check that the alert happened as
expected.

R=thestig@chromium.org

Review URL: https://codereview.chromium.org/960663002