Remove unused IDS_ constants from pdfium.
Along the way, I noticed that one of these is used in a header file
that isn't included anywhere and describes no actual code (JS_Console.h).

Also add missing header guards, and IWYU to resources.h

Also move a static function from a header to the new resources.cpp file.

Also fix the grammar on some of the error messages. Most of these
appear to be JS error messages destined for a JS error handler or (someday)
a console (which can remain in english), and oddly, some appear to be about
failed form validation, which smells user-facing.

Also check that the message makes sense for the point the C++ code
wants to use it, and reword accordingly.

As it turns out, these currently don't make it back to JS due to the
unimplemented JS_Error() function; this is an enabling step towards
implementing it.

R=thestig@chromium.org

Review URL: https://codereview.chromium.org/955273003