Make conversion between CPDF_Action and its dictionary explicit.
Precursor to taking a second shot at cleaning up the FPDF_*
APIs.  A FPDF_Action is a CPDF_Dictionary, and a CPDF_Action
is a structure holding a FPDF_Action.  This goes against the
convention that FPDF_ types get cast to CPDF_* types, so we
want to make it clear where objects are getting constructed,
etc.

Also tidy fpdf_actionhandler.cpp because it bugs me.

R=thestig@chromium.org

Review URL: https://codereview.chromium.org/984773002