Fix the pdfium component build.
It has a component with no object files in it. When compiled as a shared library, this fails.

This changes the target to a source set which can handle empty sources. It also changes the other static libraries in pdfium to source_sets (it should be very rare to need real static libraries).

Runs "gn format" over the BUILD files.

Renames the arg bundle_freetype to pdfium_bundle_freetype which makes more sense in the context of a larger build like Chrome.

R=jam@chromium.org

Review URL: https://codereview.chromium.org/985543002