Make conversions for CPDF_Link explicit.
Precursor to taking a second shot at cleaning up the FPDF_*
APIs.  A FPDF_LINK is a CPDF_Dictionary, and a CPDF_Link
is a structure holding a FPDF_LINK.  This goes against the
convention that FPDF_ types get cast to CPDF_* types, so we
want to make it clear where objects are getting constructed.

R=thestig@chromium.org

Review URL: https://codereview.chromium.org/985503005