Merge to XFA: Fix a bug that JPX images can't be shown
In the process of opj_end_decompress, it will return fail when the end of coding stream is reached. However it returns true in the same scenario implemented in openJPEG. So the final solution is from openJPEG. Return true when the end of coding stream is reached.

BUG=452671
R=tsepez@chromium.org

Review URL: https://codereview.chromium.org/990683002