Fix a problem that JP2 image is not displayed because index color space is used
There are two issues in this bug.

One is that JP2 image is not displayed because it aborts loading Jpx bitmap when the number of components in color space is different with that one in JPX images. I found that the number of components in color space isn't updated after it's initialized. For index color space, the component shall inherit from its base color space.

The second issue is that displayed color is not correct after I fixed the first issue. The root cause is that sRGB is used in JPX image, it doesn't need to map from index to RGB again.

BUG=464215
R=tsepez@chromium.org

Review URL: https://codereview.chromium.org/1009513003