Speculative fix for breakage in pdfum DEPS roll
Speculative because we still don't have pdfium buildbots for mac/win
on the pdfium waterfall.  Nonetheless, the two build errors reported
in https://codereview.chromium.org/989213003/ seem to be a default
statement in a switch fully covering all of an enum constant, and
the use of <cstdint> instead of <stdint.h> on mac.

R=thestig@chromium.org

Review URL: https://codereview.chromium.org/1011803003