Pull in specific revision of PDF corpus via DEPS.
There are two reasons for doing this. The first is that adding files
to this repository doesn't trigger a rebuild on the bot, so the next
unrelated pdfium change will be the one to hit any issues with the
unversioned corpus.

The second is that we may need a specific version of the expected output
files (.pngs) to go with a specific pdfium build.

R=thestig@chromium.org

Review URL: https://codereview.chromium.org/1019783003