Corpus roll to pick up the 2015-04-28 Foxit drop.
This will undoubtedly red up the tree, as we don't have trybots. A follow-up
CL will add the suppressions required for each platform at the moment.

The new suppressions in this CL are for cases where we didn't generate an
expected result file (due to the issue in fx/FRC_3.5_part1/Introduction.txt).

R=thestig@chromium.org

Review URL: https://codereview.chromium.org/1111213005