Create top-level public/ header directory.
These are the only files that embedders of PDFium should be including.
They are entirely self-contained, and compile cleanly against -Wall so
as to not offend the code that may include them.

Having done this, we can see that chromium is pulling in two additional
files from the fpdfsdk/include/pdfwindow directory, which is not guaranteed
to work.

A few files are renamed, adding an "_" to make the names consistent.
The exception is fpdfview, which is doc'd as such in the doc.

Naturally, paths will need updating in a handful of files in chrome
when this rolls in.

BUG=pdfium:154
R=thestig@chromium.org

Review URL: https://codereview.chromium.org/1135913002