Set pointers in internal fields directly instead of wrapping them
The internal fields are scanned by the garbage collector, so they can't contain arbitrary data. However, aligned pointers are supported by the V8 GC, so the V8
API allows for setting a pointer directly instead of wrapping it in an External
container.

Not only is this faster, but it's also required for the new v8::Global API which
I'm going to update to in a follow-up patch.

R=tsepez@chromium.org
BUG=

Review URL: https://codereview.chromium.org/1139853003