Add missing comma to third_party.gyp
A reference to nonstd_unique_ptr.h was added with
https://codereview.chromium.org/1091283002 but a comma is missing
after it. This causes this string to be concatenated with the next,
leading to broken sources. This went unnoticed by the build due to
the fact that the affected sources are both headers, but they should
be properly separated.

R=brucedawson@chromium.org

Review URL: https://codereview.chromium.org/1156663006