Use bundled freetype for the pdfium_test binary.
But use platform freetype for library itself according to the rules
for the platform.

This should greatly reduce per-platform diffs in the corpus tests, but
requires that the corpus be rolled at the same time.

When this rolls into chromium, its src/BUILD.gn will need to be updated
to say third_party:fx_freetype instead of third_party:freetype.

R=jam@chromium.org

Review URL: https://codereview.chromium.org/1267493005 .