Kill IFX_FontMapper
The design envisioned two kinds of font mappers: internal
(of type CFX_FontMapper), and external, both of which
implement a common interface.  But there isn't any means of
making an external one, so the interface is redundant.

IFX_FontMapper also appears in an apple-specific struct,
but was never non-null or used in any way.

The CFX_FontMapper class provides its own mechanism to
include platform-specific features, as it turns out.

R=thestig@chromium.org

Review URL: https://codereview.chromium.org/1259123005 .