Revert "Revert "Merge to master: contention over isolate data slots""
This reverts commit 3b4382a847b5a7439a3107512dbe54c317108579.
The difference between this CL and the one that failed is
fxjs_v8.cpp:271. In master, we pass the runtime information
as:
  v8::isolate -> v8::Context -> FXJS Runtime,

but in XFA:
  V8::Isolate -> PerIsolate struct -> FXJS Runtime.

The master way is more correct, in that FXJS_Runtime is 1:1
with v8 contexts and many:1 (in theory) with isolates.
It looks like the XFA branch missed a patch along the way.
I'll do that next.

Having made this change, the only data in the per-isolate
struct will be the ptr array (on master); it will also
include the XFA context (on XFA). I've kept the struct on
master for the sake of similarity.

R=thestig@chromium.org

Review URL: https://codereview.chromium.org/1365733003 .