Rename IFXJS_Runtime and IFXJS_Context to IJS_.
Nothing but sed.

FXJS is a layer that makes it easier to define objects in V8, but has
no knowledge of PDF-specific native objects.  It could in theory be used
to implement other sets of native objects.

JS is the layer that implements PDF-specific native objects on top of
FXJS.

Therefore, the classes used to interface to JS should be named using
IJS_. IFXJS_ is reserved for someday adding better API for FXJS iteslf.

R=thestig@chromium.org

Review URL: https://codereview.chromium.org/1386173002 .