XFA: Fix unittests and embeddertests crashers
The unittests had a case where we added an abort() call to
the code, but left the case.  And one of the expected results
was not updated with the code.

The fpdfview.cpp got broken by a recent change, and was not
caught due to all the void* returns.  Also, the tests now
clean up the page automatically.

R=thestig@chromium.org

Review URL: https://codereview.chromium.org/1403373006 .