Revert "Make m_pShadingObj a CPDF_Stream instead of CPDF_Object."
There appear to be a bunch of corpus tests which fail with this change such
as FAILURE: shading1.pdf

This reverts commit 85361b227ad6786d2aeef8409b79a8d077a26ee9.

Make m_pShadingObj a CPDF_Stream instead of CPDF_Object.

This object is required to be a stream and was being converted as such. With
the new type checking this caused us to pass a nullptr where previously we'd
have, incorrectly, cast a CPDF_Dictionary to a CPDF_Stream.

This CL changes the m_pShadingObj to always be a CPDF_Stream. Then, we never
go down the bad code path because we check if m_pShadingObj is nullptr earlier
and bail out.

BUG=chromium:547706
TBR=tsepez@chromium.org

Review URL: https://codereview.chromium.org/1414283003 .