Check if document was opened during testing.
When we call OpenDocument we fail to check if the document was actually opened.
Currently we return true in all cases (assuming we read the file). This CL
updates the code to check if the document was opened and return false if not.

I've updated several tests to check for FALSE instead of TRUE. I verified the
documents in fact don't open with my local (non-PDFium) PDF reader.

BUG=pdfium:223
R=tsepez@chromium.org

Review URL: https://codereview.chromium.org/1417893007 .