XFA: remove unsafe exif parsing code
Fortunately, this could only be called with a null buffer,
so none of unchecked lengths could be used.  The remaining
use of the CFX_/IFX_DIBAttributeEx class is as a table, so
put one directly in the CFX_DIBAttribute.

Fix a "register" warning along the way.

R=dsinclair@chromium.org

Review URL: https://codereview.chromium.org/1425983002 .