Fix incorrect CPDFSDK_PageView::CountAnnots(). (try 2)
The original XFA version was correct, and the master version here is
wrong. The two versions are now in sync, but incorrect.
So we need to fix this here and then merge to XFA again.

Also fix what are now incorrect uses of CountAnnots() and do some
cleanups.

BUG=543049
R=tsepez@chromium.org

Review URL: https://codereview.chromium.org/1413973011 .