Remove state from RebuildCrossRef state-machine.
The state '12' is only used once. This CL folds the contents of state 12 back
into the place where we set our state to 12.

This works because all state 12 does is decrement the loop counter so we process
the same character again and move us to state 0.

R=thestig@chromium.org

Review URL: https://codereview.chromium.org/1430643007 .