Get rid of most uses of CFX_PtrArray.
I didn't go whole hog and replace these with std::vector,
but in the mean time, it is silly to cast a typedef for
a template instantiated against void* when we can just
instantiate the template against the actual type.

The ones that remain are actual heterogeneous arrays with
wacky casting.

R=thestig@chromium.org

Review URL: https://codereview.chromium.org/1518593002 .