Bugs in CJS_PublicMethods::ParseNumber().
Fix the bugs by removing ParseNumber() entirely.

For PDFium's JavaScript bindings, we want to get out of the
numeric conversion business and inflict that on V8 as
possible, avoiding platform-specific issue in strtod().

For other uses, there is a FX_atof() which is similarly
buggy, but we can consolidate the use.  Add an overloaded
FX_atof() to handle wide strings more simply.

BUG=pdfium:361
R=jochen@chromium.org

Review URL: https://codereview.chromium.org/1586203006 .