Rework progressive render loop.
Migrating this off of CFX_PrtList depends on first making
this saner.  The tricky part is checking whether new
objects have been added to a partially-parsed layer.  The
old code used to do this by keeping track of the last
element in the list, and seeing if it changes. It's far
simpler to keep track of the last element processed, and
just keep looking for more objects after it with each pass.

Fold together some duplicated blocks along the way.
Remove unused EstimateProgress() method.

R=thestig@chromium.org

Review URL: https://codereview.chromium.org/1699443002 .