Split CPDF_PageObjectHolder off from CPDF_PageObjectList
Eventually, we're going to expose an iterator over
CPDF_PageObjectList that we don't want to be inherited by
the CPDF_PageObjectHolder sub-classes: page and form.

Also, the operations that the object holder performs dealing
with inquiring about masks and such seem beyond the scope of
what a list would provide.  Hence the "Holder" name.

R=thestig@chromium.org

Review URL: https://codereview.chromium.org/1701073002 .