Replace CPDF_Rect and CPDF_Point with CFX types.
Too many rectangle types.  Remove CPDF_Rect #define, and
make an equivalent point type in CFX, but stop short of
consolidating the N different forms of rectangles in CFX.

Also banish PDF_ClipFloat to the one .cpp file that needs it,
since it was in the vicinity of the .h file changes.

R=thestig@chromium.org

Review URL: https://codereview.chromium.org/1740923003 .