Structs shouldn't inherit from CFX_Target.
There are some assumptions in the code that CAPS_NAMED
structs are POD, and can be memcpy'd, malloc'd, etc.
CFX_Target adds a virtual dtor, which is not what is
desired for many of these, so resolve by making each
a CAPS_Named class or by removing the inheritence.

In the process XFA_TextPiece was found to be malloc'd
despite having a ctor.  Move this to |new|, noting that
the object is never freed (and goes out of existence
when the allocator's pool is destroyed).

BUG=pdfium:432
R=dsinclair@chromium.org

Review URL: https://codereview.chromium.org/1780753002 .