FPDF_PageDelete must delete XFA pages as well.
Currently, it is only deleting the CPDF_ resources, which
are wrapped by XFA objects in an XFA build. Hence, if a page
is deleted and then re-inserted, we get the old contents. In
print preview, chromium first inserts blank pages and then
replaces them later on, causing the associated bug.

BUG=594111
R=dsinclair@chromium.org

Review URL: https://codereview.chromium.org/1804163002 .