Support the device font cache
Reuse the Skia typeface on sucessive text draw calls.
This reduces the SKP size by 100x for some documents.

Note that this does not use a smart pointer for the
Skia typeface object. The downside of doing so is that
it requires all clients that include fx_font.h to also
have access to Skia.

In this specific case, it is preferable to have a
forward declared class to isolate Skia from the rest of PDFium.

R=dsinclair,tsepez@chromium.org

Review URL: https://codereview.chromium.org/1837113004