Rename CFX_WideStringC::raw_str() to c_str()
No functional change intended.

This difference in naming is standing in the way of consolidating
some of the string code between Wide and Byte strings. The Wide
code wants to call raw_str() in exactly the same spots that the
Byte code calls c_str().  This makes sense, because in both places we
get a character type back, and not a uint*_t type.  If WideStringC
had a raw_str() method, it would have to return uint32_t or similar.

R=dsinclair@chromium.org

Review URL: https://codereview.chromium.org/1863593004